+++ lib/src/export.dart
@@ -75,6 +75,7 @@ export './heatmap/heatmap_chart.dart';
 export './heatmap/heatmap_data.dart';
 export './pie/pie_chart.dart';
 export './radar/radar_chart.dart';
+export './gauge/gauge_chart.dart';
 export './treemap/treemap_chart.dart';
 export './treemap/treemap_data.dart';
 export './series/series_axis.dart'